+++ CHANGELOG.md
@@ -1,5 +1,28 @@
-v4.X.X (XXXX 2023)
+v4.16.0 (May 2024)
+  - Add support for Additional Websites as nodes
   - Parse inline code, not just code blocks
+
+v4.15.0 (December 2024)
+  - No changes
+
+v4.14.0 (October 2024)
+  - No changes
+
+v4.13.0 (July 2024)
+ - Add extrainformation as an available evidence field
+
+v4.12.0 (May 2024)
+ - Migrate integration to use Mappings Manager
+ - Update Dradis links in README
+
+v4.11.0 (January 2024)
+  - No changes
+
+v4.10.0 (September 2023)
+  - Update gemspec links
+
+v4.9.0 (June 2023)
+  - No changes
 
 v4.8.0 (April 2023)
   - No changes